refactor(architecture): Icon and Logo
refactor(architecture): Icon and Logo
