--- src/assets/loginPage.less
@@ -1,51 +1,0 @@
-.loginPage {
-  display: flex;
-  width: 100%;
-  height: 100vh;
-}
-
-.left {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  width: 50%;
-  height: 100%;
-  background-image: linear-gradient(90deg, #5CC04A 0%, #8FDD3D 100%);
-
-  .logo {
-    width: 80%;
-  }
-
-  .login-divier {
-    color: #FFFFFF;
-    min-width: 10%;
-    max-width: 60%;
-    width: 50%;
-    height: 3px;
-    margin: 10% 0 10% 0;
-    border-radius: 8px;
-  }
-
-  .description {
-    max-width: 40%;
-    font-size: 2vw;
-    text-align: center;
-    color: #FFFFFF;
-  }
-}
-
-.right {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  width: 50%;
-  height: 100%;
-  background-color: #FFFFFF;
-
-  .inputBox {
-    margin: 50% 0 50% 0;
-  }
-}
-
