+++ src/pages/Login.tsx
@@ -1,5 +1,5 @@
 import '../assets/Style/Login/loginPage.less';
-import {ReactComponent as Logo} from "../assets/Terradia_white.svg"
+import {ReactComponent as Logo} from "../assets/Logo/Terradia_white.svg"
 import React from "react";
 import {Divider} from "antd";
 import LoginForm from "../components/Authentication/Login/LoginForm";
